Also delete staked balances when reloading staking

+++ modules/pool/lib/staking/fantom/master-chef-staking.service.ts
@@ -8,6 +8,7 @@ import { formatFixed } from '@ethersproject/bignumber';
 import { getContractAt } from '../../../../web3/contract';
 import ERC20Abi from '../../../../web3//abi/ERC20.json';
 import { BigNumber } from 'ethers';
+import { PrismaPoolStakingType } from '@prisma/client';
 
 const FARM_EMISSIONS_PERCENT = 0.872;
 
@@ -92,10 +93,15 @@ export class MasterChefStakingService implements PoolStakingService {
         await prismaBulkExecuteOperations(operations, true);
     }
 
-    public async reloadStakingForAllPools() {
-        await prisma.prismaPoolStakingMasterChefFarmRewarder.deleteMany({});
-        await prisma.prismaPoolStakingMasterChefFarm.deleteMany({});
-        await prisma.prismaPoolStaking.deleteMany({ where: { type: 'MASTER_CHEF' } });
-        await this.syncStakingForPools();
+    public async reloadStakingForAllPools(stakingTypes: PrismaPoolStakingType[]) {
+        if (stakingTypes.includes('MASTER_CHEF')) {
+            await prisma.prismaUserStakedBalance.deleteMany({
+                where: { OR: [{ staking: { type: 'MASTER_CHEF' } }, { staking: { type: 'FRESH_BEETS' } }] },
+            });
+            await prisma.prismaPoolStakingMasterChefFarmRewarder.deleteMany({});
+            await prisma.prismaPoolStakingMasterChefFarm.deleteMany({});
+            await prisma.prismaPoolStaking.deleteMany({ where: { type: 'MASTER_CHEF' } });
+            await this.syncStakingForPools();
+        }
     }
 }
